#697: Address review comments.  Reorganise status build/status helper. Remove debugging.
