+++ wps-common/src/main/java/au/org/aodn/aws/wps/status/ExecuteStatusBuilder.java
@@ -1,13 +1,21 @@
 package au.org.aodn.aws.wps.status;
 
+import net.opengis.ows._1.CodeType;
 import net.opengis.wps._1_0.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.xml.datatype.DatatypeConfigurationException;
+import java.io.BufferedInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.HashMap;
 
+import static au.org.aodn.aws.wps.status.StatusHelper.getS3ExternalURL;
+
 public class ExecuteStatusBuilder {
+
+    static int CHUNK_SIZE = 512;
 
     private String location;
     String jobId;
@@ -25,6 +33,15 @@ public class ExecuteStatusBuilder {
         return this.location;
 
     }
+
+
+    public static final String getStatusDocument(String s3Bucket, String statusFilename, String jobId, EnumStatus jobStatus, String message, String messageCode, HashMap<String, String> outputsHrefs)
+    {
+        String statusLocation = getS3ExternalURL(s3Bucket, jobId + "/" + statusFilename);
+        ExecuteStatusBuilder statusBuilder = new ExecuteStatusBuilder(statusLocation, jobId);
+        return statusBuilder.createResponseDocument(jobStatus, message, messageCode, outputsHrefs);
+    }
+
 
     public String createResponseDocument(EnumStatus jobStatus)
     {
@@ -65,11 +82,11 @@ public class ExecuteStatusBuilder {
         }
         else if (jobStatus==EnumStatus.STARTED)
         {
-            status.setProcessStarted(StatusHelper.getProcessStartedType("Job " + jobId + " is currently running", new Integer(0)));
+            status.setProcessStarted(getProcessStartedType("Job " + jobId + " is currently running", new Integer(0)));
         }
         else if (jobStatus==EnumStatus.PAUSED)
         {
-            status.setProcessPaused(StatusHelper.getProcessStartedType("Job " + jobId + " is currently paused", new Integer(0)));
+            status.setProcessPaused(getProcessStartedType("Job " + jobId + " is currently paused", new Integer(0)));
         }
         else if (jobStatus==EnumStatus.SUCCEEDED)
         {
@@ -81,13 +98,13 @@ public class ExecuteStatusBuilder {
                 {
                     String href = outputs.get(currentKey);
                     LOGGER.info("OUTPUT [" + currentKey + "]=[" + href + "]");
-                    StatusHelper.addExecuteOutputReference(response, currentKey, href);
+                    addOutputToResponse(response, currentKey, href);
                 }
             }
         }
         else if (jobStatus==EnumStatus.FAILED)
         {
-            status.setProcessFailed(StatusHelper.getProcessFailedType(failedMessage, failedCode));
+            status.setProcessFailed(getProcessFailedType(failedMessage, failedCode));
         }
 
         response.setStatus(status);
@@ -100,6 +117,79 @@ public class ExecuteStatusBuilder {
     public static void main(String[] args) {
         System.out.println();
         ExecuteStatusBuilder e = new ExecuteStatusBuilder();
+    }
+
+
+    public static ProcessFailedType getProcessFailedType(String message, String code)
+    {
+        ProcessFailedType failed = new ProcessFailedType();
+        failed.setExceptionReport(StatusHelper.getExceptionReport(message, code));
+        return failed;
+    }
+
+
+    public static ProcessStartedType getProcessStartedType(String message, Integer percentComplete)
+    {
+        ProcessStartedType started = new ProcessStartedType();
+        started.setValue(message);
+        started.setPercentCompleted(percentComplete);
+        return started;
+    }
+
+
+    public static void addOutputToResponse(ExecuteResponse response, String outputIdentifier, String outputHref)
+    {
+        OutputDataType output = new OutputDataType();
+
+        OutputReferenceType outputReference = new OutputReferenceType();
+        outputReference.setHref(outputHref);
+        output.setReference(outputReference);
+
+        CodeType outputIdentifierCode = new CodeType();
+        outputIdentifierCode.setValue(outputIdentifier);
+        output.setIdentifier(outputIdentifierCode);
+        if(response.getProcessOutputs() == null) {
+            ExecuteResponse.ProcessOutputs outputs = new ExecuteResponse.ProcessOutputs();
+            response.setProcessOutputs(outputs);
+        }
+        response.getProcessOutputs().getOutput().add(output);
+    }
+
+
+    public static void addOutputToResponse(ExecuteResponse response, String outputIdentifier, InputStream inStream)
+    {
+        OutputDataType output = new OutputDataType();
+
+        LiteralDataType literalData = new LiteralDataType();
+        literalData.setDataType("??");
+        literalData.setUom("??");
+        if(inStream != null)
+        {
+            StringBuilder strBuilder = new StringBuilder();
+            byte[] chunk = new byte[CHUNK_SIZE];
+            BufferedInputStream bufferedStream = new BufferedInputStream(inStream);
+            try {
+                int bytesRead = bufferedStream.read(chunk);
+                while (bytesRead != -1) {
+                    strBuilder.append(new String(chunk, 0, bytesRead));
+                }
+            }
+            catch(IOException ioex)
+            {
+                //  Error out
+                throw new RuntimeException("Unable to add literal output to response : " + ioex.getMessage(), ioex);
+            }
+
+            //  Contents of the stream
+            literalData.setValue(strBuilder.toString());
+        }
+        DataType data = new DataType();
+        data.setLiteralData(literalData);
+        output.setData(data);
+        CodeType outputIdentifierCode = new CodeType();
+        outputIdentifierCode.setValue(outputIdentifier);
+        output.setIdentifier(outputIdentifierCode);
+        response.getProcessOutputs().getOutput().add(output);
     }
 }
